requirements.txt
discord.py==1.5.1
numpy==1.19.2
Pillow==8.3.2
psycopg2==2.8.6
requests==2.23.0
python-dotenv==0.19.1
boto3==1.19.7